==> ryan_r_files/03_multi_xgboost.qmd <==
---
title: "Multivariate XGBoost Model"
subtitle: "Predicting COVID-19 Deaths (STAT 390)" 
author: "Ryan Nguyen"

format:
  html:
    toc: true
    embed-resources: true
    link-external-newwindow: true
    code-fold: true
    
output:
  html_document:
    code_folding: hide

execute:
  warning: false

from: markdown+emoji 
---

## 1. Import Libraries & Set Seed
```{r}
library(forecast)
library(tseries)
library(tidyverse)
library(tidymodels)
library(kableExtra)
library(modeltime)
library(prophet)
library(timetk)
library(Metrics)
library(zoo)
library(MLmetrics)
library(xgboost)

set.seed(123)

```


## 2. Pull Data
```{r}
selected_deaths <- read_csv("data/finaldata_selectedfeatures.csv")
```



## 3. Train Test Split
```{r}

selected_deaths <- fill(selected_deaths, everything())

selected_deaths <- replace(selected_deaths, is.na(selected_deaths), 0)
 
naniar::miss_var_summary(selected_deaths)
 
selected_deaths <- selected_deaths %>% 
   mutate(covid_19_deaths = case_when((covid_19_deaths == 0) ~ 1,
                        .default = covid_19_deaths))  #%>% 
 # select(-c(starts_with("deaths_"), "dayofweek", "quarter", "month", "dayofyear", "dayofmonth",
            #"weekofyear"))

#selected_deaths <- selected_deaths %>% 
 # mutate(covid_19_deaths = log(covid_19_deaths))

# Splitting into testing and training by year
train_data  <- selected_deaths %>% 
  filter(year != 2023)

test_data <- selected_deaths %>%
  filter(year == 2023)

# Converting data into xg boost matrices
train_matrix <- train_data %>% 
  select(-c(date, covid_19_deaths)) %>% 
  as.matrix()

train_matrix2 <- train_matrix %>% 
  xgb.DMatrix

test_matrix2 <- test_data %>% 
  select(-c(date, covid_19_deaths)) %>% 
  as.matrix() 

test_matrix <- test_data %>% 
  select(-c(date, covid_19_deaths)) %>% 
  as.matrix() %>% 
  xgb.DMatrix()

# creating target vector
targets <- train_data$covid_19_deaths
```

## 4. Control Grid
```{r}
xgb_trcontrol <- caret::trainControl(
   method = "cv", 
   number = 5, # 5 folds
   allowParallel = TRUE, 
   verboseIter = FALSE, 
   returnData = FALSE
)

xgb_grid <- base::expand.grid(
   list(
    nrounds = c(100, 200),
    max_depth = c(10, 15, 20), # maximum depth of a tree
    colsample_bytree = seq(0.5), # subsample ratio of columns when construction each tree
    eta = 0.01, # learning rate
    gamma = 0, # minimum loss reduction
    min_child_weight = 1,  # minimum sum of instance weight (hessian) needed ina child
    subsample = 1 # subsample ratio of the training instances
))
```


```{r eval = FALSE}
# wlist = list(train = train_matrix2, test = test_matrix)

xgb_model <- caret::train(
   train_matrix, targets,
   trControl = xgb_trcontrol,
   tuneGrid = xgb_grid,
   #early_stopping_rounds = 10,
   #watchlist = wlist,
   method = "xgbTree",
   nthread = 1
)

save(xgb_model, file = "results/model_xgboost.rda")
```

## 5. Check Best Model
```{r}
load("results/model_xgboost.rda")

xgb_model$bestTune
```

## 6. Perform Forecast
```{r}
xgb_pred <- xgb_model %>% stats::predict(test_data)
```

## 7. Model Plot
```{r}
# prediction on a train set
fitted <- xgb_model %>%
    stats::predict(train_data)

xgtrain_predictions <- cbind(fitted, train_data) %>% 
  mutate(index = seq(1:6974))
xgtest_predictions <-cbind(xgb_pred, test_data) %>% 
  rename(fitted = xgb_pred) %>% 
  mutate(index = seq(1:332))
# xgfull_pred <- rbind(xgtrain_predictions, xgtest_predictions)

```

```{r}

xgplot <- xgtest_predictions %>% 
  ggplot() +
  geom_line(aes(x = index, y = fitted, color = "orange")) +
  #geom_point(aes(x = index, y = fitted, color = "blue"), alpha = 0.2) +
  geom_line(aes(x = index, y = covid_19_deaths, color = "dodgerblue")) +
  #geom_point(aes(x = index, y = covid_19_deaths, color = "red"), alpha = 0.2) +
  scale_color_manual(name="Value", values = c("orange", "dodgerblue"), labels = c("Predicted", "Actual")) +
  labs(title = "XGBoost Forecast", 
       y = "Deaths") + 
  coord_fixed(ratio = .3) +
  theme_bw()

xgplot
```


## 7. Model Performance
```{r}
# Unlogging the values
# xgb_pred2 <- as.data.frame(xgb_pred) %>%
#   mutate(xgb_pred = exp(xgb_pred))
# 
# test_data2 <- test_data %>%
#   select(covid_19_deaths) %>%
#   mutate(covid_19_deaths = exp(covid_19_deaths))

# Check MAE value
xgboost_mae <- mae(test_data$covid_19_deaths, xgb_pred)
print(paste("The MAE for the XGBoost model is", xgboost_mae))

# Check MASE value
xgboost_mase <- mase(test_data$covid_19_deaths, xgb_pred)
print(paste("The MASE for the XGBoost model is", xgboost_mase))
```

```{r eval = FALSE}
# Saving everything to compare results
save(xgboost_mae, xgboost_mase, xgtrain_predictions, xgtest_predictions, xgplot,
     file = "results/multi_xgboost.rda")

```
