==> ryan_r_files/04_uni_results.qmd <==
---
title: "Univariate Results"
subtitle: "Predicting COVID-19 Deaths (STAT 390)" 
author: "Ryan Nguyen"

format:
  html:
    toc: true
    embed-resources: true
    link-external-newwindow: true
    code-fold: true
    
output:
  html_document:
    code_folding: hide

execute:
  warning: false

from: markdown+emoji 
---

```{r, echo = FALSE, results = 'hide', message = FALSE}
library(patchwork)
library(tidyverse)
library(kableExtra)
library(knitr)

load("results/uni_sarima.rda")
load("results/uni_autoarima.rda")
load("results/uni_arima.rda")
# load("results/uni_prophet.rda")
load("results/univariate_prophet.rda")
```

## Plots

### East Region
```{r}
#| label: East Plots
east_initial_arima_plot +
  east_tuned_arima_plot +
  east_initial_sarima_plot + 
  east_tuned_sarima_plot + 
  east_auto_plot + 
  east_best_proph_test + 
  plot_layout(ncol = 2)
```

### Midwest Region
```{r}
#| label: Midwest Plots
midwest_initial_arima_plot +
  midwest_tuned_arima_plot +
  midwest_initial_sarima_plot + 
  midwest_tuned_sarima_plot + 
  midwest_auto_plot + 
  midwest_best_proph_test +
  plot_layout(ncol = 2)
```

### South Region
```{r}
#| label: South Plots
south_initial_arima_plot +
  south_tuned_arima_plot +
  south_initial_sarima_plot + 
  south_tuned_sarima_plot + 
  south_auto_plot + 
  south_best_proph_test +
  plot_layout(ncol = 2)
```

### West Region 
```{r}
#| label: West Plots
west_initial_arima_plot +
  west_tuned_arima_plot +
  west_initial_sarima_plot +
  west_tuned_sarima_plot + 
  west_auto_plot + 
  west_best_proph_test +
  plot_layout(ncol = 2)
```


## Model Accuracy

### East Region
```{r}
tribble(
  ~model, ~mae, ~mase,
  #--|--|----
  "initial arima", east_initial_arima_mae, east_initial_arima_mase,
  "tuned arima", east_tuned_arima_mae, east_tuned_arima_mase,
  "initial sarima", east_initial_sarima_mae, east_initial_sarima_mase,
  "tuned sarima", east_tuned_sarima_mae, east_tuned_sarima_mase,
  "auto arima", east_auto_mae, east_auto_mase,
  "prophet", east_season_prophet_mae, east_season_prophet_mase
) %>% 
  kbl() %>% 
  kable_styling()
```

### Midwest Region
```{r}
tribble(
  ~model, ~mae, ~mase,
  #--|--|----
  "initial arima", midwest_initial_arima_mae, midwest_initial_arima_mase,
  "tuned arima", midwest_tuned_arima_mae, midwest_tuned_arima_mase,
  "initial sarima", midwest_initial_sarima_mae, midwest_initial_sarima_mase,
  "tuned sarima", midwest_tuned_sarima_mae, midwest_tuned_sarima_mase,
  "auto arima", midwest_auto_mae, midwest_auto_mase,
  "prophet", midwest_best_prophet_mae, midwest_best_prophet_mase
) %>% 
  kbl() %>% 
  kable_styling()
```

### South Region
```{r}
tribble(
  ~model, ~mae, ~mase,
  #--|--|----
  "initial arima", south_initial_arima_mae, south_initial_arima_mase,
  "tuned arima", south_tuned_arima_mae, south_tuned_arima_mase,
  "initial sarima", south_initial_sarima_mae, south_initial_sarima_mase,
  "tuned sarima", south_tuned_sarima_mae, south_tuned_sarima_mase,
  "auto arima", south_auto_mae, south_auto_mase,
  "prophet", south_best_prophet_mae, south_best_prophet_mase
) %>% 
  kbl() %>% 
  kable_styling()
```

### West Region
```{r}
tribble(
  ~model, ~mae, ~mase,
  #--|--|----
  "initial arima", west_initial_arima_mae, west_initial_arima_mase,
  "tuned arima", west_tuned_arima_mae, west_tuned_arima_mase,
  "initial sarima", west_initial_sarima_mae, west_initial_sarima_mase,
  "tuned sarima", west_tuned_sarima_mae, west_tuned_sarima_mase,
  "auto arima", west_auto_mae, west_auto_mase,
  "prophet", west_best_prophet_mae, west_best_prophet_mase
) %>% 
  kbl() %>% 
  kable_styling()


```
